+++ src/css/components/help-menu.less
@@ -14,7 +14,7 @@
     color: @white;
   }
 
-  &-dropdown.popover.bottom {
+  &-dropdown.popover.bottom, &-dropdown.popover.default {
     margin: 0;
 
     .content {
@@ -23,7 +23,7 @@
       min-width: auto;
       padding: 0;
       left: 100%;
-      transform: translate(-110%, 50%);
+      transform: translate(-90%, 10%);
 
       &:before, &:after {
         display: none;
